# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "注册成功，token 已写入 httpOnly cookie"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "登录成功，token 已写入 httpOnly cookie"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/qq/sign-in": {
      "post": {
        "operationId": "QqAuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QqAccessTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "QQ 已绑定，本站登录成功"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/qq/bind": {
      "post": {
        "operationId": "QqAuthController_bind",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QqAccessTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "QQ 已绑定到当前账号"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "mylog3 API",
    "description": "后端接口文档 — 前端通过 openapi-typescript 生成类型",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SignUpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 20,
            "description": "用户昵称，全局唯一",
            "example": "alice"
          },
          "pswd": {
            "type": "string",
            "minLength": 6,
            "description": "登录密码，明文传入，后端 hash 后入库"
          },
          "captchaId": {
            "type": "string",
            "description": "图形验证码 id（/captcha/create 返回的 id）"
          },
          "captcha": {
            "type": "string",
            "description": "图形验证码用户输入"
          }
        },
        "required": [
          "name",
          "pswd",
          "captchaId",
          "captcha"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "用户昵称",
            "example": "alice"
          },
          "pswd": {
            "type": "string",
            "description": "登录密码，明文传入"
          }
        },
        "required": [
          "name",
          "pswd"
        ]
      },
      "QqAccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "QQ JS SDK 返回的 Access Token"
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
    }
  }
}